Improved microcontroller assignation and deletion.
Up to now, the only way to assign or delete a microcontroller over an
object was specify the full path of the object.
Now is possible to assign and delete a microcontroller specifying only
the container or a pseudo-folder, and the assignation/deletion request
will be applied to all objects below this path.